Created the CollectableDatabase

+++ app/src/main/java/com/qrhunter/Collectable.java
@@ -6,17 +6,16 @@ import android.widget.ImageView;
 
 public class Collectable {
     private String id;
-    private int score;
+    private Long score;
     private Bitmap photo;
     private Pair<Double, Double> location = new Pair<>(0.0, 0.0);      // lat, long
 
-
-    public int getScore() {return score;}
+    public Long getScore() {return score;}
     public String getId() {return id;}
     public Bitmap getPhoto() {return photo;}
     public Pair<Double, Double> getLocation() {return location;}
 
-    public void setScore(int new_score) {
+    public void setScore(Long new_score) {
         assert(new_score >= 0);
         score = new_score;
     }
